feat: move get assets
