+++ src/libs/cdn/src/storage/state/stable.rs
@@ -8,6 +8,7 @@ use junobuild_shared::types::core::Blob;
 use junobuild_storage::types::state::FullPath;
 use junobuild_storage::types::store::{Asset, AssetEncoding};
 use std::borrow::Cow;
+use std::ops::RangeBounds;
 
 pub fn get_asset(
     cdn_stable: &impl CdnStableStrategy,
@@ -45,6 +46,38 @@ fn get_content_chunks_impl(
     let key: ContentChunkKey =
         deserialize_from_bytes(Cow::Owned(encoding.content_chunks[chunk_index].clone()));
     content_chunks.get(&key)
+}
+
+pub fn get_assets(
+    cdn_stable: &impl CdnStableStrategy,
+    proposal_id: &ProposalId,
+) -> Vec<(AssetKey, Asset)> {
+    cdn_stable.with_assets(|assets| get_assets_impl(proposal_id, assets))
+}
+
+fn get_assets_impl(
+    proposal_id: &ProposalId,
+    proposal_assets: &AssetsStable,
+) -> Vec<(AssetKey, Asset)> {
+    proposal_assets
+        .range(filter_assets_range(proposal_id))
+        .collect()
+}
+
+fn filter_assets_range(proposal_id: &ProposalId) -> impl RangeBounds<AssetKey> {
+    let start_key = AssetKey {
+        proposal_id: *proposal_id,
+        collection: "".to_string(),
+        full_path: "".to_string(),
+    };
+
+    let end_key = AssetKey {
+        proposal_id: *proposal_id + 1,
+        collection: "".to_string(),
+        full_path: "".to_string(),
+    };
+
+    start_key..end_key
 }
 
 fn stable_asset_key(
